gcp renaming > renamed "gcp" to "google cloud"
