+++ CONTRIBUTING.md
@@ -1,7 +1,7 @@
 # Contributing
 
 Welcome to the HiveMQ Community!
-Glad to see your interest in contributing to HiveMQ GCP PubSub Extension Hello World Customization.
+Glad to see your interest in contributing to HiveMQ Google Cloud PubSub Extension Hello World Customization.
 Please checkout our [Contribution Guide](https://github.com/hivemq/hivemq-community/blob/master/CONTRIBUTING.adoc) to make sure your contribution will be accepted by the HiveMQ team.
 
 For information on how the HiveMQ Community is organized and how contributions will be accepted please have a look at our [HiveMQ Community Repo](https://github.com/hivemq/hivemq-community).
