+++ gradle.properties
@@ -2,7 +2,7 @@ version=4.9.0-SNAPSHOT
 #
 # main dependencies
 #
-hivemq-gcp-pubsub-sdk.version=4.9.0,rc-1
+hivemq-google-cloud-pubsub-sdk.version=4.9.0-SNAPSHOT
 #
 # test dependencies
 #
