+++ src/test/java/com/hivemq/extensions/google/cloud/pubsub/customizations/helloworld/PubSubToMqttHelloWorldTransformerTest.java
@@ -1,14 +1,29 @@
-package com.hivemq.extensions.gcp.pubsub.customizations.helloworld;
+/*
+ * Copyright 2022-present HiveMQ GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hivemq.extensions.google.cloud.pubsub.customizations.helloworld;
 
 import com.codahale.metrics.MetricRegistry;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.packets.general.Qos;
 import com.hivemq.extension.sdk.api.services.builder.PublishBuilder;
 import com.hivemq.extension.sdk.api.services.publish.Publish;
-import com.hivemq.extensions.gcp.pubsub.api.model.InboundPubSubMessage;
-import com.hivemq.extensions.gcp.pubsub.api.transformers.PubSubToMqttInitInput;
-import com.hivemq.extensions.gcp.pubsub.api.transformers.PubSubToMqttInput;
-import com.hivemq.extensions.gcp.pubsub.api.transformers.PubSubToMqttOutput;
+import com.hivemq.extensions.google.cloud.pubsub.api.model.InboundPubSubMessage;
+import com.hivemq.extensions.google.cloud.pubsub.api.transformers.PubSubToMqttInitInput;
+import com.hivemq.extensions.google.cloud.pubsub.api.transformers.PubSubToMqttInput;
+import com.hivemq.extensions.google.cloud.pubsub.api.transformers.PubSubToMqttOutput;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -18,7 +33,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicReference;
 
-import static com.hivemq.extensions.gcp.pubsub.customizations.helloworld.PubSubToMqttHelloWorldTransformer.MISSING_DATA_COUNTER_NAME;
+import static com.hivemq.extensions.google.cloud.pubsub.customizations.helloworld.PubSubToMqttHelloWorldTransformer.MISSING_DATA_COUNTER_NAME;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
