+++ src/test/java/com/hivemq/extensions/google/cloud/pubsub/customizations/helloworld/TestPubSubConnection.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2022-present HiveMQ GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hivemq.extensions.google.cloud.pubsub.customizations.helloworld;
+
+import com.hivemq.extension.sdk.api.annotations.NotNull;
+import com.hivemq.extensions.google.cloud.pubsub.api.model.PubSubConnection;
+
+/**
+ * @author Florian Limpöck
+ * @since 4.9.0
+ */
+class TestPubSubConnection implements PubSubConnection {
+
+    @Override
+    public @NotNull
+    String getId() {
+        return "my-connection";
+    }
+
+    @Override
+    public @NotNull
+    String getProjectId() {
+        return "my-project";
+    }
+}
